chore(e2e): adds localStorage setup and teardown step (#3030)
Adds a gherkin step to add values to localStorage:

```gherkin
Given the localStorage
  """
    key:
        this:
           will: end
           up: json
           encoded: true     
  """
```
The root key will be the localStorage key, and the rest will get JSON
stringified and used for the value. So the above will be:

```
| Key | Value                                                   |
| key | {"this": { will: "end", "up": "json", "encoded": true}} | 
```

We never use anything but this form of JSON encoding, and we shouldn't
ever use anything else either.

---

I also combined `visit.ts` here. We used to have this separate for
reasons, but we no longer need it to be separate. (ln:103 down is just
the contents of the deleted file)

Signed-off-by: John Cowen <john.cowen@konghq.com>