chore: adds XLayout component (#3092)
Adds a `XLayout` component for laying out multiple components.

`XLayout` currently comes with two `types` of layout: `stack` and
`separated`

- `stack` is the same as what `div class="stack"` does, renders the
contents vertically with a gap in-between each item.
- `separated` is the same as the layout we use when rendering multiple
badges and renders the contents inline/horizontally with a gap
in-between each item.

---

We currently use `<div class="stack">` (and others div/class things) to
lay out certain components. This uses `<XLayout type="stack">` to do
exactly the same thing. But formalises it in the fact that is no longer
just a div and a class, its a very simple component that you can use
inline in your template with auto-complete-able layouts.

Another advantage is that we are making all our layouts consistent. For
example, and specifically added in this PR, the margin between our
tables and pagination is now shared with other "stacks". The margin is
no longer a property (under the ownership) of a `KTable`, it's now owned
by an XLayout. This way we can guarantee that the margin used for
separating other components is exactly the same as the margin that
separates a KTable/KPagination, previously we would have to
guess/investigate what the margin was that was being used inside of
KTable.

For `separated`, `TagList` currently uses this rule as CSS, but I'd like
to be able to use it to render lists of `TargetRef` components, `Label`
components, or maybe just Badge components. The idea is we can now
re-use this simple layout via a composable layout component.

Lastly, I haven't refactored everything to use this as yet, I can do
that in a follow up PR.

---------

Signed-off-by: John Cowen <john.cowen@konghq.com>