chore: remove unnecessary env vars (#3049)
Back in the mists of time we used a plugin in order to interpolate
variables into our `index.html` file. We need this as in 'production'
kuma itself uses our `index.html` as a template and interpolates
variables into it.

Since then vite was upgraded to v5 which included functionality to
interpolate variables, but _only_ shell environment variables. This led
us to drop the plugin for variable interpolation and use environment
variables instead. The only thing is this meant we also started adding
single line stringified JSON into environment variables.

This has always sat a little strange to me, these variables don't need
to be shell environment variables and the stringified JSON is awkward.

Additionally lately we've realised we want to use a variable from the
`package.json` file (`"version": ...`) to prevent duplication of this
value in several places. (see
https://github.com/kumahq/kuma-gui/pull/2967)

Because of this last point, I decided it was a good time to revisit
this.

---

This PR aims to replicate exactly what the kuma binary does. We use JS
to search and replace the `golang` style `{{.BaseGuiPath}}` variables
with the actual variables. ~This is achieved using a relatively simple
almost single line "inline vite plugin".~ As we replicate what kuma
does, it means we don't have to do anything special anymore for
'production', we just ship the template as is, then only in
'development' we interpolate variables.

This also removes the need for the shell environment variables and the
awkward stringified single-line JSON, and means we can reference the
`package.json#version` via JS where we need to.

Lastly, I removed some fallback JSON that we had for a case that I don't
think I've ever seen happen (an un-parseable HTML file), and really, if
the HTML is un-parsable we probably don't want anyone using the
resulting GUI as its likely to be very broken and/or unadvisable to use.

Overall this PR simplifies things quite a bit.

---

## Post initial approval additions
(https://github.com/kumahq/kuma-gui/pull/3049/commits/873ab088cac57795241586590443dc6d0c3a0806)

There has unfortunately been quite a bit of back and forth here, partly
because I needed to check that this was all going to work elsewhere. I
also noticed another area where we do similar template interpolating
(our preview server that runs a production build during e2e tests).

Due to those realisations I:

- Created a `vite.plugins.ts` file to house the things I wanted to reuse
across vite configs (this new file will soon be moved to a centralized
config area)
- Added 2 vite plugins here.
- A plugin that works with `configureServer` _and_
`configurePreviewServer` for local dev and for running a production
build for e2e testing
- A plugin that works for our docs site, that unfortunately needs to use
a slightly different approach as we can't provide an index.html to
vitepress to use, therefore we sneakily inject HTML into the index.html
that vitepress uses.

All in all I'm happier that everything is now using the same approach,
_and_ we do _nothing_ for production i.e. there is nothing to break.

---------

Signed-off-by: John Cowen <john.cowen@konghq.com>