feature(routing): adds route config walking, plus several walkers (#3151)
This PR introduces the concept of "route config walkers" to our
application.

These walkers allow us to decorate individual route configs in a
centralized place. They purposefully mutate the config rather than
return the a copy of the config. I've kept the code for this in one
place for now, but once I'm totally sure of the API I'll probably
re-shuffle things slightly to use our service container labels. I've
added code comments/TODOs to this effect in the PR.

---

The walkers/decorators I added here:

### `$routeName`

Currently we have/use `<RouteView name="the-name-of-the-route">` i.e. we
hard code the name of the route associated with this file into the
RouteView's name. This means we can't then reuse, or 'mount', the same
route component file at another URL and at the very least its error/typo
prone having to write the thing in two places. I actually wanted to do
something like this when I added `RouteView` orignally but my Vue chops
weren't good enough to think of a way to do it 😅

For example, in relation to
https://github.com/kumahq/kuma-gui/issues/2921, I'd like to 'mount' the
`subscription/` module in various places: `zones/`, `zone-ingresses/`,
`zone-egresses/` and `data-planes/`. I can't hardcode in
`zone-cps-subscription-summary-view` as the name for the route when its
under zones, because that will be incorrect when I re-use it for say
`zone-ingress-subscription-summary-view`. We need to know the name of
the route dynamically.

`addRouteName` adds a dynamically assigned `$routeName` global var to
every single route component template rendering context:

- whilst following vue pseudo-standard `$` prefixing for global vars
- whilst avoiding the need for the template author to configure a prop
all the time (by using the global)
- I don't think there is a way via typescript to say "this var exists in
a route component, but not in any other types of components" so even
though this is always set for route components, I set it to possibly
undefined so that its correct if someone accidentally thinks its
available in non-route components. This also explains why I'm ok using
`$routeName!` (with `!`) here. I want to make sure `:name` is required,
but I'm totally sure that in a route component `$routeName` is set.

### meta.module

For quite a while now our application is split into several
microservices/micro-frontends/modules (or whatever we want to call
them). It would be handy sometimes to know the module that a route
belongs to. For example in relation to
https://github.com/kumahq/kuma-gui/issues/2430. We would eventually like
to avoid having to type
`t('name-of-the-module.name-of-the-route.title')` and just write
`t('title')` from within route components. If `RouteView::t` knows which
module and route it is in, we can avoid having to repeat the
prefix/namespace 10s of times throughout the route component, hopefully
making it less of a chore to use i18n for anything not super essential.

I would guess there might be other places where we would use this.

`addMeta` lets you add a `meta.module` to the very top route in a
modules route configuration. This then flows down to every single child
route in that module so we don't have to hardcode that ourselves. We add
it once, and it flows down. Note: I haven't added any top-level
`meta.module` for this to use as yet.

### meta.path

I also think it will be useful at some point to have a full path for a
routing configuration as well as just the name.

`addPath` automatically compiles a `meta.path` on every single route
config containing the full path of a route. For example a route called
`zone-ingress-subscription-summary-view` would have a `meta.path`
similar to
`zone-cp-detail-view.zone-ingress-detail-view.zone-ingress-subscription-summary-view`.

I hope at some point we'll be able to reduce some more redundant
namespacing in the future and expand on the concept of 'mounting'
modules at several different places.

For example, instead of a name of
`zone-ingress-subscription-summary-view` I could have a route path of
`zone-cp.zone-ingress.subscription.summary`. I might also want to link
to simply `.summary.config` (the `config` of whereever I'm mounted)
instead of `zone-cp.zone-ingress.subscription.summary.config`

---

Lastly I only need `$routeName` right now for
https://github.com/kumahq/kuma-gui/issues/2921. Both `meta.module` and
`meta.path` are features for other further work, but it was useful to
add several functions to check the API (for example I didn't realise at
first I might need access to `parent: RouteRecordRaw`)

---------

Signed-off-by: John Cowen <john.cowen@konghq.com>