feat: introduce hostname-generators (#3160)
This change introduces the views for `HostnameGenerator`s, including
`ListView`, `SummaryView` and `DetailView`. With the next iterations we
will add more information to the views.
Currently there is no navigation element on the home page that brings
the user to the views. For now the path `/hostname-generators` has to be
entered manually.

Closes #2956

---------

Signed-off-by: schogges <moritz.fleck@konghq.com>