chore: adds new @kumahq/config module to house shareable config (#3184)
Creates and moves a bunch of tooling config files into a new
`@kumahq/config` module.

1. eslint config
2. stylelint config
3. Cypress config
4. a CI script that we use for sharding our tests to feed to Cypress so
we can run them in parallel.

Couple of other things:

1. I noticed a `docgen.config` file that we no longer use/need so I
deleted that here also.
2. I tried to be intentional with the new `@kumahq/config/package.json`
dependencies. i.e. depend on the things it uses unless the host app also
uses it (in which case use a `peerDependency`). I'm not totally sure if
it matters in the grand scheme of things, but even if a tiny detail does
matter it will be good to be intentional/correct about what we want to
convey here. As a consequence of this a load of `dev: true`s where
removed form our `package-lock.json`. I don't think this matters too
much but thought I'd mention in case.

---------

Signed-off-by: John Cowen <john.cowen@konghq.com>