feat(kuma-gui/control-planes): add ControlPlaneActionGroup with link to HostnameGenerators (#3189)
This change introduces a new `XActionGroup` to the root overview
(`ControlPlaneDetailView`) and includes a link to the
`HostnameGenerators`.

There was also a little refactoring of the `XActionGroup` necessary. In order to have a responsive design of that component, the dropdown now grows with the content until a certain point. Very long labels break at that point. This will also help at the point when introducing another language in terms of `i18n`.

Closes #3174

---------

Signed-off-by: schogges <moritz.fleck@konghq.com>