feat(data-planes/data-plane-summary-view): add config (#3209)
This change introduces a tab navigation to the `DataPlaneSummaryView`
offering an `Overview` and `Config` tab. The reason for the separation
into tabs was because the summary view could get very crowded already.
The `Overview` tab shows everything that was previously shown on the
summary view. The `Config` tab shows the config, which is based on some
metadata and `dataplane.network` data of a `DataPlaneOverview` item
(similar to what was done in #3202 and #3203).

The `DataPlaneSummaryView` is used in multiple places and I've added an
optional prefix for the route names.

In order to avoid path name collision, the dataplane summary routes use a `summary-` prefix for the tabs. In terms of consistency this is not ideal, but to keep the deeplinking to the respective tabs on the details view, this was necessary.

Also, I replaced the custom type `DataplaneOverview` in the data layer
and infer it now from the return value of
`DataPlaneOverview.fromObject`.

Part of #2922

---------

Signed-off-by: schogges <moritz.fleck@konghq.com>