feat: default service hosts to period-delimited (#3222)
DNS is created in the mesh for `_`-delimited and `.`-delimited
hostnames. This PR makes the `.`-delimited the default (as it is the
only one in conformance with the RFC). This is a QOL improvement for
folks copying/pasting from the UI, and both versions should work
identically. The API returns `_`-delimited so in the case anyone was
using any kind of automation, this shouldn't break them anyway.

fixes #3098

Signed-off-by: John Harris <john@johnharris.io>