refactor(app/navigation): split sidebar navigation, move diagnostics to configuration (#3213)
This change splits up the sidebar navigation into a top nav including
the nav links to the most important routes. The new bottom nav offers a
place for other routes. Both navigation lists are separated by a visual
separator. This ensures that the user finds all nav links to top level
resources in the same area of the UI. Both nav lists are only split
visually and remain within a single `nav` element.
To further beautify the nav links I've added icons to each of them.

I've moved and renamed everything from `diagnostics` to `configuration`.
This makes it more clear that this part of the UI shows the CP
configuration.

---------

Signed-off-by: schogges <moritz.fleck@konghq.com>