feature(dataplanes): add filtered XDS config to drawers (#3186)
Adds XDS configuration drawers to inbounds and outbounds so show only
the listeners/clusters relevant to that inbound/outbound.

Closes https://github.com/kumahq/kuma-gui/issues/3182
Closes https://github.com/kumahq/kuma-gui/issues/3181

You can run this against a local cluster by setting a cookie like the
following:


note: I removed an annoying ESLint rule here
https://eslint.org/docs/latest/rules/multiline-ternary#when-not-to-use-it
. We don't have any conventions in regards to ternaries, do whatever
works for you. Generally you've decided to use a ternary because the
format you've personally chosen reads nicer - furthermore its pointless
having a eslint rule to force wrapping of something that is designed for
single lines. Not only that but the eslint rule we had makes these super
hard to read.


---------

Signed-off-by: John Cowen <john.cowen@konghq.com>