chore: lint package.json versions to only allow caret semvers (#3235)
Honestly I'm not a huge fan of linting, a lot of rules that people use
are either unnecessary or over-zealous. Not only that but it always
means depending on lots of things like `eslint`, which you then have to
maintain (dependabot PRs etc etc), just for the sake of not caring too
much about things like having alphabetically ordered import statements.

Now I've got that off my chest 😅 , I do agree that there are sometimes
use cases for specific rules to prevent mistakes that can lead to
genuine bugs.

This use case:

Over the past couple of years we've had several issues with duplicate
dependencies appearing in our package lock file. They are easy to miss
in PR reviews, and a lot of the time people don't realise that hainvg
duplicate dependencies can cause very hard to spot issues (in our case
having a duplicate install of kongponents has led to CSS issues that are
not caught by CI as we don't have CSS regression tests).

The reason we've ended up with a duplicate install of a dependency is
because somebody somewhere perhaps mistakenly has used an exact version
to specify a package version in their consumable package. This goes for
_at least_ `dependencies` and `peerDependencies`, which effect consuming
applications.

I think there is no argument/discussion about the need to use range
constraints rather than exact constraints in consumable packages (as
oppose to consuming applications). In fact we should really go the extra
mile even if there is a problem and the temptation to fix a version
exactly in a consumable package. (Just so its said, IMO, we should also
only use range constraints for consuming applications unless its
absolutely necessary - but we are not here to talk about that specific
use case today.)

This PR:

- Installs `"eslint-plugin-json-schema-validator"` so we can validate
JSON files (I know I know another dependency...)
- Uses a minimal custom json schema to validate our package.json file,
only allowing caret versions i.e. `^1.0.0`
- Adds an exception for `@kuma/config` to only allow `*`
- Comments in examples of ignoring our caret rule so we have an
escape-hatch/ignore for our rule.
- Includes the rule in our `make lint/js` **for '@kumahq/kuma-gui'
only**. We don't yet apply all out linting to our entire workspace, but
we will soon.

None of this is an issue in our repository as yet, (we always use `^`
versions because as a team we've agreed to so we don't need a linter),
but we have flip-flopped in the past. Also in our case we have
comparatively little to manage and a small enough team to get away with
no hard-defined lint rules, others might not be so lucky.

Kinda unrelated but interesting:

- Also adds a rule to disallow `engines`. Eventually I would like to use
this to have a single source of truth for our `engines` values in our
workspace `package.json`, whilst we can't include files from
`@kumahq/config` into our root package.json file, at least we can lint
it to make sure it has specific values using a single source of truth
linting rule.

Lastly:

I started with this https://github.com/tclindner/npm-package-json-lint,
but decided against it seeing as its a standalone tool (we already
need/require eslint so best if we can continue to use that rather than
add another tool). I also found it to be fairly rigid, i.e. whilst there
is a rule to enforce carets on `devDependencies` and `dependencies`
there strangely isn't one for `peerDependencies` and it looks like they
don't support adding custom rules. I then thought about looking for a
more generic solution where I could define my own rules - a generic JSON
schema linter fits that problem, plus you never know as its generic we
might find other uses for it.

---------

Signed-off-by: John Cowen <john.cowen@konghq.com>
Signed-off-by: John Cowen <johncowen@users.noreply.github.com>
Co-authored-by: Mo <24444046+schogges@users.noreply.github.com>