chore(dataplanes): move envoy related sources to connections (#3265)
We have a new task to add some of the traffic visualization and envoy
debug information over in Zones for ZoneProxies
(https://github.com/kumahq/kuma-gui/issues/2976).

Zone Proxies (Ingresses/Egresses) are more or less just Dataplanes, or
maybe better said just Envoy instances.

It's maybe better to think of a Zone Proxy as a Kuma thing that is
essentially an Envoy Instance and similarly a Dataplane as a Kuma thing
that is essentially an Envoy Instance.

Our current modules related to this are:

- `zone-ingresses`
- `zone-egresses`
- `data-planes`

and

- `connections`

Currently `data-planes` depends on `connections` for the traffic
functionality related to `inbounds` and `outbounds` (a.k.a `'bounds`,
a.k.a. `connections`)

You can kinda see that `zones-'gresses` should also depend on
`connections`, but I think its further than that.

---

There is a lot of older code from before we made `connections` that is
currently in `data-planes`. This is functionality/views for `XDS
Configuration`, `Stats` and `Clusters`.

We also have almost identical code in `zones-*gresses` for
functionality/views `XDS Configuration`, `Stats` and `Clusters`.

This functionality/views for `XDS Configuration`, `Stats` and `Clusters`
should therefore be part of `connections` (a.k.a. "the envoy related
stuff") and we reuse that for all modules that need it (similar to how
we did `subscriptions`)

This gives us something like:

- `zone-ingresses` depends on `connections` 
- `zone-egresses` depends on `connections` 
- `data-planes` depends on `connections` 

Instead of:

- `zone-ingresses` has its own stuff to display Envoy things 
- `zone-egresses` has its own stuff to display Envoy things 
- `data-planes` depends on `connections` 

This PR moves any DataSource `sources` related to `stats`, `xds` or
`clusters` over to `connections`.

I also tried moving the Views in this PR, but it was so frustratingly
difficult and finickity to do because of the way we "flatly" structure
and "overly" name our route files in a very inconsistent and error prone
way that I pulled it back out and plan to PR that separately. I'll have
more to say on this at some point and will probably link back here.

---

Working towards https://github.com/kumahq/kuma-gui/issues/2976

Signed-off-by: John Cowen <john.cowen@konghq.com>