feat(gateways/builtin-gateway-list-view): add summary drawer with config (#3284)
As part of the efforts to show the config "everywhere" I've now also
added a summary drawer for the `builtin` gateway. I could only add it
for the `builtin` gateway because this can be configured via
`MeshGateway`. For the `delegated` gateway this is not possible because
the config is part of other resources like `Dataplane` (refer to
[kuma.io/managing-ingress-traffic/delegated](https://kuma.io/docs/2.9.x/using-mesh/managing-ingress-traffic/delegated/)).

Part of #2922

---------

Signed-off-by: schogges <moritz.fleck@konghq.com>