refactor(common/resource-status): prefer undefined over null (#3301)
Generally we want to avoid handling `null` anywhere else than the data
layer. In the data layer we make sure to transform all values to
something that we can safely access in the view layer (e.g. no
conditional chaining required). As a last resort we prefer `undefined`
over `null` if something is just not set.
To meet these requirements I've refactored this component. It's
important to mention that we specifically need the case of `online`
being `undefined`, because there are cases in which we only want to show
the `total`.

---------

Signed-off-by: schogges <moritz.fleck@konghq.com>