refactor: add and use XAboutCard component in detail views (#3300)
These changes introduce a new global component `XAboutCard` that is used
to improve the layout of the formerly used `KCard` in the detail views
"About" section. It is based on the `AppAboutSection` exported by
`@kong-ui-public/app-layout`. The layout in the content slot changed to
a horizontal layout with badges (`XBadge`) instead of the column based
layout for a general improvement of readability. All `XAboutSection`
show the date times of creation and potential modification of the
resource.


![image](https://github.com/user-attachments/assets/6e657857-464a-482e-9b92-66ebd2a562b8)

For better visual separation of the `ResourceStatus` entries in the
`ControlPlaneStatus` I have introduced a new class
`.columns-with-borders` which works similarly to `.stack-with-borders`
but for a horizontal layout.

![Screenshot 2024-12-13 at 18 28
17](https://github.com/user-attachments/assets/7242b18a-c348-4671-8cbc-8430c2831733)

Furthermore I have replaced several occurrences of `class="column"` or
`class="stack"` in the files I touched with `XLayout` and replaced
static text with i18n.

---------

Signed-off-by: schogges <moritz.fleck@konghq.com>