feat(zones): Add zone-ingress traffic view (#3296)
I've been working on a new traffic visualisation for zone-ingresses
similar to our visualisation that we have for Dataplanes.

I'm made some progress, but felt like it was a good time to put some
work up for review, partly to preview progress and also to PR some
scattered/tiny but important refactors.

Things here:

- Moved the exclusion filter that we apply when parsing `/stats` deeper
into the data layer, this bit will apply everywhere we parse stats.
- Added a new `/connections/stats/...` source modelled on the one we
have for dataplanes. Note: I plan to move all the sources we have in
this file to use a `/connections/...` prefix (the same name as the
module)
- When building up to do this work, I went over our existing Dataplane
viz, mainly to refresh my memory and stare at it a bit to figure out how
to re-use or re-implement this for zones. Whilst doing that I made some
very small amends. This should only be upgrading components, and maybe a
bit of a rename/re-nest. There should be no functionality changes in
this file.
- Finally, new work in the Zone Ingress Detail page. Note: The cards
don't currently link to drawers, that will be the next PR.

---------

Signed-off-by: John Cowen <john.cowen@konghq.com>